ログイン機能、削除機能の修正

+++ app/controllers/tasks_controller.rb
@@ -1,5 +1,5 @@
 class TasksController < ApplicationController
-
+  before_action :authenticate_user!, only: [:new, :edit, :destory]
   def index
    @tasks = Task.order(created_at: :desc)
   end
@@ -14,6 +14,7 @@ class TasksController < ApplicationController
 
   def create
    @task = Task.create(task_params)
+   @task.user_id = current_user.id
    if @task.save
     redirect_to root_path
    else
@@ -23,6 +24,7 @@ class TasksController < ApplicationController
 
   def edit
     @task = Task.find(params[:id])
+    redirect_to action: :index unless user_signed_in? && current_user.id == @task.user_id
   end
 
   def update
@@ -32,14 +34,19 @@ class TasksController < ApplicationController
   end
 
   def destroy
-    task = Task.find(params[:id])
-    task.destroy
-    redirect_to tasks_path
+    @task = Task.find(params[:id])
+    if user_signed_in? && current_user.id == @task.user_id && @task.destroy
+      redirect_to tasks_path
+    else
+      render :show
+    end
   end
 
  private
+
    def task_params
-    params.require(:task).permit(:title, :sentence, :year_id, :month_id, :day_id, :dayofweek_id, :weather_id)
+    params.require(:task).permit(:title, :sentence, :year_id, :month_id, :day_id, :dayofweek_id, :weather_id).merge(user_id: current_user.id)
    end
 
 end
+
